newsvendor-llm-experiment/pyproject.toml
[tool.poetry]
name = "newsvendor-llm-experiment"
version = "0.5.0"
description = "LLM Agent Newsvendor Negotiation Experiment"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
ollama = "^0.3.0"
pandas = "^2.1.0"
numpy = "^1.24.0"
scipy = "^1.11.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
plotly = "^5.15.0"
scikit-learn = "^1.3.0"
pydantic = "^2.4.0"
pyyaml = "^6.0"
asyncio = "^3.4.3"
aiofiles = "^23.2.0"
loguru = "^0.7.0"
tqdm = "^4.66.0"
click = "^8.1.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
psutil = "^5.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"
pre-commit = "^3.4.0"
jupyter = "^1.0.0"

[tool.poetry.scripts]
newsvendor-experiment = "src.experiments.run_full_experiment:main"
newsvendor-validate = "src.experiments.run_validation_suite:main"
newsvendor-analyze = "src.experiments.analyze_results:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true